Refactored patch js files
