+++ db-controller.js
@@ -11,7 +11,10 @@ import { index, idNegotiation, generateSlugId, remove } from './controllers/util
 import { create, query, id } from './controllers/crud.js'
 import { searchAsWords, searchAsPhrase } from './controllers/search.js'
 import { deleteObj } from './controllers/delete.js'
-import { putUpdate, patchUpdate, patchSet, patchUnset, overwrite } from './controllers/update.js'
+import { putUpdate, overwrite } from './controllers/update.js'
+import { patchUpdate } from './controllers/patchUpdate.js'
+import { patchSet } from './controllers/patchSet.js'
+import { patchUnset } from './controllers/patchUnset.js'
 import { bulkCreate, bulkUpdate } from './controllers/bulk.js'
 import { since, history, idHeadRequest, queryHeadRequest, sinceHeadRequest, historyHeadRequest } from './controllers/history.js'
 import { release } from './controllers/release.js'
