+++ routes/__tests__/set.test.js
@@ -28,15 +28,15 @@ it("'/set' route functions", async () => {
     .set('Content-Type', 'application/json; charset=utf-8')
     .then(resp => resp)
     .catch(err => err)
-      expect(response.header.location).toBe(response.body["@id"])
-      expect(response.headers["location"]).not.toBe(`${process.env.RERUM_ID_PREFIX}11111`)
-      expect(response.statusCode).toBe(200)
-      expect(response.body._id).toBeUndefined()
-      expect(response.body["test_set"]).toBe(unique)
-      expect(response.headers["content-length"]).toBeTruthy()
-      expect(response.headers["content-type"]).toBeTruthy()
-      expect(response.headers["date"]).toBeTruthy()
-      expect(response.headers["etag"]).toBeTruthy()
-      expect(response.headers["allow"]).toBeTruthy()
-      expect(response.headers["link"]).toBeTruthy()
+  expect(response.header.location).toBe(response.body["@id"])
+  expect(response.headers["location"]).not.toBe(`${process.env.RERUM_ID_PREFIX}11111`)
+  expect(response.statusCode).toBe(200)
+  expect(response.body._id).toBeUndefined()
+  expect(response.body["test_set"]).toBe(unique)
+  expect(response.headers["content-length"]).toBeTruthy()
+  expect(response.headers["content-type"]).toBeTruthy()
+  expect(response.headers["date"]).toBeTruthy()
+  expect(response.headers["etag"]).toBeTruthy()
+  expect(response.headers["allow"]).toBeTruthy()
+  expect(response.headers["link"]).toBeTruthy()
 })
