+++ routes/__tests__/unset.test.js
@@ -27,15 +27,15 @@ it("'/unset' route functions", async () => {
     .set('Content-Type', 'application/json; charset=utf-8')
     .then(resp => resp)
     .catch(err => err)
-    expect(response.header.location).toBe(response.body["@id"])
-    expect(response.statusCode).toBe(200)
-    expect(response.body._id).toBeUndefined()
-    expect(response.body.hasOwnProperty("test_obj")).toBe(false)
-    expect(response.headers["content-length"]).toBeTruthy()
-    expect(response.headers["content-type"]).toBeTruthy()
-    expect(response.headers["date"]).toBeTruthy()
-    expect(response.headers["etag"]).toBeTruthy()
-    expect(response.headers["allow"]).toBeTruthy()
-    expect(response.headers["link"]).toBeTruthy()
+  expect(response.header.location).toBe(response.body["@id"])
+  expect(response.statusCode).toBe(200)
+  expect(response.body._id).toBeUndefined()
+  expect(response.body.hasOwnProperty("test_obj")).toBe(false)
+  expect(response.headers["content-length"]).toBeTruthy()
+  expect(response.headers["content-type"]).toBeTruthy()
+  expect(response.headers["date"]).toBeTruthy()
+  expect(response.headers["etag"]).toBeTruthy()
+  expect(response.headers["allow"]).toBeTruthy()
+  expect(response.headers["link"]).toBeTruthy()
 })
